chg: [stats] fixed some routing issues and simplify code in csskp.urls. separate data accessible only for authenticated admins and data which do not require authentication.
