+++ admin/views.py
@@ -1,16 +1,17 @@
 # -*- coding: utf-8 -*-
 
 import sys
+import json
 from django.shortcuts import render
-from django.db.models import Count
 from django.http import JsonResponse, HttpResponseRedirect
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
-from django.conf.global_settings import LANGUAGES
 from utils.utils import exec_cmd, exec_cmd_no_wait
+from stats.stats_logic import get_finished_surveys_list
 from survey.reporthelper import calculateResult
 from survey.lib.utils import export_survey
 from survey.models import SurveyUser
+from stats.forms import DatesLimitForm
 
 
 @login_required
@@ -18,32 +19,6 @@ def export_survey_json(request):
     """Returns the survey (questions, answers, recommendations and sections) in JSON."""
     result = export_survey()
     return JsonResponse(result, safe=False)
-
-
-@login_required
-def survey_status_count(request):
-    """Returns the count for the SurveyUser status property."""
-    result = SurveyUser.objects.values("status").annotate(count=Count("status"))
-    status = {1: "In progress", 2: "Under reviews", 3: "Finished"}
-    return JsonResponse(
-        {status[item["status"]]: item["count"] for item in result.all()}
-    )
-
-
-@login_required
-def survey_language_count(request):
-    """Returns the count for the SurveyUser chosen_lang property."""
-    result = SurveyUser.objects.values("chosen_lang").annotate(
-        count=Count("chosen_lang")
-    )
-    return JsonResponse(
-        {
-            [lang for lang in LANGUAGES if lang[0] == item["chosen_lang"]][0][1]: item[
-                "count"
-            ]
-            for item in result.all()
-        }
-    )
 
 
 @login_required
@@ -60,6 +35,28 @@ def site_stats(request):
         "survey_results": survey_results,
     }
     return render(request, "admin/site_stats.html", context=context)
+
+
+@login_required
+def results_stats(request):
+    if request.method == "POST":
+        dates_limit_form = DatesLimitForm(data=request.POST)
+        if not dates_limit_form.is_valid():
+            return None
+        start_date = dates_limit_form.cleaned_data["start_date"]
+        end_date = dates_limit_form.cleaned_data["end_date"]
+    else:
+        dates_limit_form = DatesLimitForm()
+        start_date = dates_limit_form.fields["start_date"].initial
+        end_date = dates_limit_form.fields["end_date"].initial
+
+    result = get_finished_surveys_list(start_date, end_date)
+    context = {
+        "dates_limit_form": dates_limit_form,
+        "surveys_users_results": json.dumps(result, indent=2, sort_keys=True)
+    }
+
+    return render(request, "admin/results_stats.html", context=context)
 
 
 @login_required
