+++ stats/views.py
@@ -1,25 +1,28 @@
 # -*- coding: utf-8 -*-
 
-from django.contrib.auth.decorators import login_required
+import sys
 from django.conf.global_settings import LANGUAGES
 from django.db.models import Count
 from django.shortcuts import render
-from django.http import HttpResponseRedirect, JsonResponse
-from stats.stats_logic import get_finished_surveys_list
+from django.http import JsonResponse
 from survey.models import SurveyUser
 from survey.reporthelper import calculateResult
 
 
-@login_required
 def index(request):
-    result = get_finished_surveys_list(request)
-    if result is None:
-        return HttpResponseRedirect("/admin/export/statistics/")
+    nb_finished_surveys = SurveyUser.objects.filter(status=3).count()
+    nb_surveys = SurveyUser.objects.count()
+    context = {
+        "nb_surveys": nb_surveys,
+        "nb_finished_surveys": nb_finished_surveys,
+        "python_version": "{}.{}.{}".format(*sys.version_info[:3]),
+    }
 
-    return render(request, "admin/stats.html", result)
+    return render(request, "survey/stats.html", context=context)
 
 
 def overall(request):
+    """Returns the page which will display some statistics."""
     nb_finished_surveys = SurveyUser.objects.filter(status=3).count()
     nb_surveys = SurveyUser.objects.count()
     last_surveys = SurveyUser.objects.filter(status=3).order_by("-created_at")[:10]
@@ -31,6 +34,15 @@ def overall(request):
     }
 
     return JsonResponse(result)
+
+
+def survey_status_count(request):
+    """Returns the count for the SurveyUser status property."""
+    result = SurveyUser.objects.values("status").annotate(count=Count("status"))
+    status = {1: "In progress", 2: "Under reviews", 3: "Finished"}
+    return JsonResponse(
+        {status[item["status"]]: item["count"] for item in result.all()}
+    )
 
 
 def survey_language_count(request):
