+++ templates/admin/index.html
@@ -15,7 +15,7 @@
       <h2>Statistics</h2>
       <ul class="list-group">
         <li class="list-group-item"><a href="/admin/site_stats">statistics page</a>;</li>
-        <li class="list-group-item"><a href="{% url 'stats:index' %}">export results of finished surveys</a>.</li>
+        <li class="list-group-item"><a href="/admin/results_stats">export results of finished surveys</a>.</li>
       </ul>
     </div>
   </div>
