+++ templates/admin/results_stats.html
@@ -13,7 +13,7 @@
 </div><br />
 <div class="row">
   <div class="col">
-    <form action="/admin/export/statistics/" method="post" class="dates-limit-form">
+    <form action="/admin/results_stats" method="post" class="dates-limit-form">
       <div class="form-group">
         {% csrf_token %}
         {{ dates_limit_form.media }}
