+++ templates/parts/footer_part.html
@@ -17,7 +17,7 @@
       </a>
     </li>
     <li class="nav-item">
-      <a class="nav-link" href="{{ version_url }}" title="{{ _('Version of') }} {{ custom.tool_name }}" alt="{{ _('Version of') }} {{ custom.tool_name }}"  target="_blank" rel="noopener noreferrer">
+      <a class="nav-link" href="{{ version_url }}" title="{{ _('Version of') }} {{ custom.tool_name }}" alt="{{ _('Version of') }} {{ custom.tool_name }}" target="_blank" rel="noopener noreferrer">
         {{ app_version }}
       </a>
     </li>
